Added Make, cleaned up Rake
It occurs to me that plenty of people aren't even going to have ruby
installed on their boxen, much less rake...

For now (and forever, I hope), the build process is simple enough to
duplicate in a Makefile. So why not? :-)

Accordingly, I updated the README.

I also tweaked the Rakefile a bit to take advantage of its capabilities.
Otherwise, what's the point of even having one, since I support Make? ;-)